requirements.txt
flask
flask_cors
dotenv
azure-ai-vision-imageanalysis
